+++ stories/02-Design System/04-DataEntry/Select/examples/download.tsx
@@ -52,7 +52,7 @@ const App = () => {
     const style = {width: 320};
 
     return (
-        <KitSpace direction="vertical">
+        <KitSpace direction="vertical" size={2}>
             <KitDivider orientation="left">Download compact</KitDivider>
             <KitSpace.Compact>
                 <KitSelect