+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### 🐛 Bug Fixes
 * **Select**
     * Remove unused props `suffixIcon` and adapt the documentation (@Delmotte-Vincent)
+    * Add focus state on select dropdown (@Delmotte-Vincent)
 
 ### 📚 Documentation
 * New documentation page to list available design tokens (@evoiron)